Alterando tipo de dado de Balance

+++ excContaBancaria/excContaBancaria/Account.cs
@@ -4,7 +4,7 @@
 	{
 		public int Agency { get; set; } //Agencia
 		public int Number { get; set; } //Numero da conta
-		public long Balance { get; set; } //Saldo
+		public float Balance { get; set; } //Saldo
 
 		public Account() 
 		{
